server/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Werkzeug==2.2.3
email-validator==2.0.0.post2
python-dotenv==1.0.1
gunicorn==21.2.0
selenium==4.15.2
webdriver-manager==4.0.1
